backend/pyproject.toml
[project]
name = "datareaper-backend"
version = "0.1.0"
description = "Autonomous privacy defense backend for DataReaper"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "asyncpg>=0.30.0",
  "email-validator>=2.2.0",
  "fastapi>=0.116.0",
  "greenlet>=3.1.1",
  "httpx>=0.28.1",
  "orjson>=3.10.18",
  "playwright>=1.52.0",
  "psycopg[binary]>=3.2.9",
  "pydantic>=2.11.4",
  "pydantic-settings>=2.9.1",
  "python-dotenv>=1.1.0",
  "python-multipart>=0.0.20",
  "pyyaml>=6.0.2",
  "sqlalchemy>=2.0.40",
  "structlog>=25.3.0",
  "typer>=0.15.3",
  "uvicorn[standard]>=0.34.2",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.15.0",
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0",
  "pytest-cov>=6.1.1",
  "ruff>=0.11.8",
]

[project.scripts]
datareaper = "datareaper.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
